+++ packages/accounts-controller/src/AccountsController.test.ts
@@ -67,6 +67,7 @@ const defaultState: AccountsControllerState = {
     accounts: {},
     selectedAccount: '',
   },
+  accountIdByAddress: {},
 };
 
 const mockGetKeyringByType = jest.fn();
@@ -380,6 +381,9 @@ describe('AccountsController', () => {
             },
             selectedAccount: mockSnapAccount.id,
           },
+          accountIdByAddress: {
+            [mockSnapAccount.address]: mockSnapAccount.id,
+          },
         },
         messenger,
       });
@@ -421,6 +425,9 @@ describe('AccountsController', () => {
               [mockSnapAccount.id]: mockSnapAccount,
             },
             selectedAccount: mockSnapAccount.id,
+          },
+          accountIdByAddress: {
+            [mockSnapAccount.address]: mockSnapAccount.id,
           },
         },
         messenger,
@@ -464,6 +471,9 @@ describe('AccountsController', () => {
             },
             selectedAccount: mockSnapAccount.id,
           },
+          accountIdByAddress: {
+            [mockSnapAccount.address]: mockSnapAccount.id,
+          },
         },
         messenger,
       });
@@ -506,6 +516,9 @@ describe('AccountsController', () => {
               [mockSnapAccount.id]: mockSnapAccount,
             },
             selectedAccount: mockSnapAccount.id,
+          },
+          accountIdByAddress: {
+            [mockSnapAccount.address]: mockSnapAccount.id,
           },
         },
         messenger,
@@ -552,6 +565,9 @@ describe('AccountsController', () => {
             },
             selectedAccount: mockSnapAccount.id,
           },
+          accountIdByAddress: {
+            [mockSnapAccount.address]: mockSnapAccount.id,
+          },
         },
         messenger,
       });
@@ -579,6 +595,7 @@ describe('AccountsController', () => {
             accounts: {},
             selectedAccount: '',
           },
+          accountIdByAddress: {},
         },
         messenger,
       });
@@ -619,6 +636,7 @@ describe('AccountsController', () => {
             accounts: {},
             selectedAccount: '',
           },
+          accountIdByAddress: {},
         },
         messenger,
       });
@@ -657,6 +675,7 @@ describe('AccountsController', () => {
             accounts: {},
             selectedAccount: '',
           },
+          accountIdByAddress: {},
         },
         messenger,
       });
@@ -699,6 +718,10 @@ describe('AccountsController', () => {
                 [mockAccount3.id]: mockAccount3,
               },
               selectedAccount: mockAccount.id,
+            },
+            accountIdByAddress: {
+              [mockAccount.address]: mockAccount.id,
+              [mockAccount3.address]: mockAccount3.id,
             },
           },
           messenger,
@@ -768,6 +791,10 @@ describe('AccountsController', () => {
                 [mockAccount4.id]: mockAccount4,
               },
               selectedAccount: mockAccount.id,
+            },
+            accountIdByAddress: {
+              [mockAccount.address]: mockAccount.id,
+              [mockAccount4.address]: mockAccount4.id,
             },
           },
           messenger,
@@ -844,6 +871,10 @@ describe('AccountsController', () => {
               },
               selectedAccount: mockAccount.id,
             },
+            accountIdByAddress: {
+              [mockAccount.address]: mockAccount.id,
+              [mockAccount4.address]: mockAccount4.id,
+            },
           },
           messenger,
         });
@@ -905,6 +936,9 @@ describe('AccountsController', () => {
               },
               selectedAccount: mockAccount.id,
             },
+            accountIdByAddress: {
+              [mockAccount.address]: mockAccount.id,
+            },
           },
           messenger,
         });
@@ -952,6 +986,10 @@ describe('AccountsController', () => {
                 [mockAccount2.id]: mockAccount2,
               },
               selectedAccount: mockAccount.id,
+            },
+            accountIdByAddress: {
+              [mockAccount.address]: mockAccount.id,
+              [mockAccount2.address]: mockAccount2.id,
             },
           },
           messenger,
@@ -1021,6 +1059,11 @@ describe('AccountsController', () => {
                 [mockAccount2WithCustomName.id]: mockAccount2WithCustomName,
               },
               selectedAccount: mockAccount.id,
+            },
+            accountIdByAddress: {
+              [mockAccount.address]: mockAccount.id,
+              [mockAccount2WithCustomName.address]:
+                mockAccount2WithCustomName.id,
             },
           },
           messenger,
@@ -1093,6 +1136,7 @@ describe('AccountsController', () => {
               accounts: {},
               selectedAccount: 'missing',
             },
+            accountIdByAddress: {},
           },
           messenger,
         });
@@ -1135,6 +1179,10 @@ describe('AccountsController', () => {
               },
               selectedAccount: mockAccount.id,
             },
+            accountIdByAddress: {
+              [mockAccount.address]: mockAccount.id,
+              [mockAccount3.address]: mockAccount3.id,
+            },
           },
           messenger,
         });
@@ -1169,6 +1217,9 @@ describe('AccountsController', () => {
                 [mockAccount.id]: mockAccount,
               },
               selectedAccount: mockAccount.id,
+            },
+            accountIdByAddress: {
+              [mockAccount.address]: mockAccount.id,
             },
           },
           messenger,
@@ -1237,6 +1288,10 @@ describe('AccountsController', () => {
               },
               selectedAccount: mockAccount.id,
             },
+            accountIdByAddress: {
+              [mockAccount.address]: mockAccount.id,
+              [mockAccount2.address]: mockAccount2.id,
+            },
           },
           messenger,
         });
@@ -1292,6 +1347,10 @@ describe('AccountsController', () => {
                 [mockAccount2.id]: mockAccount2,
               },
               selectedAccount: 'missing-account',
+            },
+            accountIdByAddress: {
+              [mockAccount.address]: mockAccount.id,
+              [mockAccount2.address]: mockAccount2.id,
             },
           },
           messenger,
@@ -1356,6 +1415,10 @@ describe('AccountsController', () => {
                 [mockAccount2.id]: mockAccount2WithoutLastSelected,
               },
               selectedAccount: 'missing-account',
+            },
+            accountIdByAddress: {
+              [mockAccount.address]: mockAccount.id,
+              [mockAccount2.address]: mockAccount2.id,
             },
           },
           messenger,
@@ -1429,6 +1492,10 @@ describe('AccountsController', () => {
               },
               selectedAccount: 'missing-account',
             },
+            accountIdByAddress: {
+              [mockAccount.address]: mockAccount.id,
+              [mockAccount2.address]: mockAccount2.id,
+            },
           },
           messenger,
         });
@@ -1471,6 +1538,10 @@ describe('AccountsController', () => {
                 [mockAccount3.id]: mockAccount3,
               },
               selectedAccount: mockAccount.id,
+            },
+            accountIdByAddress: {
+              [mockAccount.address]: mockAccount.id,
+              [mockAccount3.address]: mockAccount3.id,
             },
           },
           messenger,
@@ -1548,6 +1619,9 @@ describe('AccountsController', () => {
               [mockInitialAccount.id]: mockInitialAccount,
             },
             selectedAccount: mockInitialAccount.id,
+          },
+          accountIdByAddress: {
+            [mockInitialAccount.address]: mockInitialAccount.id,
           },
         },
         messenger,
@@ -1627,6 +1701,10 @@ describe('AccountsController', () => {
               },
               selectedAccount: 'unknown',
             },
+            accountIdByAddress: {
+              [mockExistingAccount1.address]: mockExistingAccount1.id,
+              [mockExistingAccount2.address]: mockExistingAccount2.id,
+            },
           },
           messenger,
         });
@@ -1669,6 +1747,7 @@ describe('AccountsController', () => {
             accounts: {},
             selectedAccount: '',
           },
+          accountIdByAddress: {},
         },
         messenger,
       });
@@ -1742,6 +1821,9 @@ describe('AccountsController', () => {
               [account.id]: account,
             },
             selectedAccount: account.id,
+          },
+          accountIdByAddress: {
+            [account.address]: account.id,
           },
         },
         messenger,
@@ -1850,6 +1932,11 @@ describe('AccountsController', () => {
               },
               selectedAccount: mockNewerEvmAccount.id,
             },
+            accountIdByAddress: {
+              [mockOlderEvmAccount.address]: mockOlderEvmAccount.id,
+              [mockNewerEvmAccount.address]: mockNewerEvmAccount.id,
+              [mockBtcAccount.address]: mockBtcAccount.id,
+            },
           },
           messenger,
         });
@@ -1875,6 +1962,10 @@ describe('AccountsController', () => {
               },
               selectedAccount: mockBtcAccount.id,
             },
+            accountIdByAddress: {
+              [mockOlderEvmAccount.address]: mockOlderEvmAccount.id,
+              [mockBtcAccount.address]: mockBtcAccount.id,
+            },
           },
           messenger,
         });
@@ -1899,6 +1990,9 @@ describe('AccountsController', () => {
                 [mockOlderEvmAccount.id]: mockOlderEvmAccount,
               },
               selectedAccount: mockOlderEvmAccount.id,
+            },
+            accountIdByAddress: {
+              [mockOlderEvmAccount.address]: mockOlderEvmAccount.id,
             },
           },
           messenger,
@@ -1999,6 +2093,7 @@ describe('AccountsController', () => {
             accounts: {},
             selectedAccount: '',
           },
+          accountIdByAddress: {},
         },
         messenger,
       });
@@ -2064,6 +2159,7 @@ describe('AccountsController', () => {
             accounts: {},
             selectedAccount: '',
           },
+          accountIdByAddress: {},
         },
         messenger,
       });
@@ -2117,6 +2213,7 @@ describe('AccountsController', () => {
             accounts: {},
             selectedAccount: '',
           },
+          accountIdByAddress: {},
         },
         messenger,
       });
@@ -2167,6 +2264,9 @@ describe('AccountsController', () => {
               [mockAccount.id]: mockAccount,
             },
             selectedAccount: mockAccount.id,
+          },
+          accountIdByAddress: {
+            [mockAccount.address]: mockAccount.id,
           },
         },
         messenger,
@@ -2238,6 +2338,7 @@ describe('AccountsController', () => {
             accounts: {},
             selectedAccount: '',
           },
+          accountIdByAddress: {},
         },
         messenger,
       });
@@ -2310,6 +2411,7 @@ describe('AccountsController', () => {
             accounts: {},
             selectedAccount: '',
           },
+          accountIdByAddress: {},
         },
         messenger,
       });
@@ -2382,6 +2484,7 @@ describe('AccountsController', () => {
             accounts: {},
             selectedAccount: '',
           },
+          accountIdByAddress: {},
         },
         messenger,
       });
@@ -2432,6 +2535,7 @@ describe('AccountsController', () => {
             accounts: {},
             selectedAccount: '',
           },
+          accountIdByAddress: {},
         },
         messenger,
       });
@@ -2531,6 +2635,10 @@ describe('AccountsController', () => {
               },
               selectedAccount: 'unknown',
             },
+            accountIdByAddress: {
+              [mockExistingAccount1.address]: mockExistingAccount1.id,
+              [mockExistingAccount2.address]: mockExistingAccount2.id,
+            },
           },
           messenger,
         });
@@ -2612,6 +2720,9 @@ describe('AccountsController', () => {
               [mockHdAccount.id]: mockHdAccount,
             },
             selectedAccount: mockHdAccount.id,
+          },
+          accountIdByAddress: {
+            [mockHdAccount.address]: mockHdAccount.id,
           },
         },
         messenger,
@@ -2700,6 +2811,9 @@ describe('AccountsController', () => {
             },
             selectedAccount: mockHdAccount.id,
           },
+          accountIdByAddress: {
+            [mockHdAccount.address]: mockHdAccount.id,
+          },
         },
         messenger,
       });
@@ -2727,6 +2841,7 @@ describe('AccountsController', () => {
             accounts: {},
             selectedAccount: '',
           },
+          accountIdByAddress: {},
         },
       });
 
@@ -2737,6 +2852,9 @@ describe('AccountsController', () => {
           },
           selectedAccount: mockAccount.id,
         },
+        accountIdByAddress: {
+          [mockAccount.address]: mockAccount.id,
+        },
       });
 
       expect(accountsController.state).toStrictEqual({
@@ -2745,6 +2863,9 @@ describe('AccountsController', () => {
             [mockAccount.id]: mockAccount,
           },
           selectedAccount: mockAccount.id,
+        },
+        accountIdByAddress: {
+          [mockAccount.address]: mockAccount.id,
         },
       });
     });
@@ -2755,6 +2876,9 @@ describe('AccountsController', () => {
           internalAccounts: {
             accounts: { [mockAccount.id]: mockAccount },
             selectedAccount: mockAccount.id,
+          },
+          accountIdByAddress: {
+            [mockAccount.address]: mockAccount.id,
           },
         },
       });
@@ -2769,6 +2893,9 @@ describe('AccountsController', () => {
           },
           selectedAccount: mockAccount.id,
         },
+        accountIdByAddress: {
+          [mockAccount.address]: mockAccount.id,
+        },
       });
     });
   });
@@ -2780,6 +2907,9 @@ describe('AccountsController', () => {
           internalAccounts: {
             accounts: { [mockAccount.id]: mockAccount },
             selectedAccount: mockAccount.id,
+          },
+          accountIdByAddress: {
+            [mockAccount.address]: mockAccount.id,
           },
         },
       });
@@ -2794,6 +2924,9 @@ describe('AccountsController', () => {
           internalAccounts: {
             accounts: { [mockAccount.id]: mockAccount },
             selectedAccount: mockAccount.id,
+          },
+          accountIdByAddress: {
+            [mockAccount.address]: mockAccount.id,
           },
         },
       });
@@ -2815,6 +2948,11 @@ describe('AccountsController', () => {
               [mockAccount3.id]: mockAccount3,
             },
             selectedAccount: mockAccount.id,
+          },
+          accountIdByAddress: {
+            [mockAccount.address]: mockAccount.id,
+            [mockAccount2.address]: mockAccount2.id,
+            [mockAccount3.address]: mockAccount3.id,
           },
         },
       });
@@ -2855,6 +2993,11 @@ describe('AccountsController', () => {
               },
               selectedAccount: lastSelectedAccount.id,
             },
+            accountIdByAddress: {
+              [mockOlderEvmAccount.address]: mockOlderEvmAccount.id,
+              [mockNewerEvmAccount.address]: mockNewerEvmAccount.id,
+              [mockBtcAccount.address]: mockBtcAccount.id,
+            },
           },
         });
 
@@ -2871,6 +3014,9 @@ describe('AccountsController', () => {
             },
             selectedAccount: mockBtcAccount.id,
           },
+          accountIdByAddress: {
+            [mockBtcAccount.address]: mockBtcAccount.id,
+          },
         },
       });
 
@@ -2886,6 +3032,7 @@ describe('AccountsController', () => {
             accounts: {},
             selectedAccount: '',
           },
+          accountIdByAddress: {},
         },
       });
 
@@ -2948,6 +3095,11 @@ describe('AccountsController', () => {
               },
               selectedAccount: selectedAccount.id,
             },
+            accountIdByAddress: {
+              [mockOlderEvmAccount.address]: mockOlderEvmAccount.id,
+              [mockNewerEvmAccount.address]: mockNewerEvmAccount.id,
+              [mockBtcAccount.address]: mockBtcAccount.id,
+            },
           },
         });
 
@@ -2973,6 +3125,11 @@ describe('AccountsController', () => {
               },
               selectedAccount: mockBtcAccount.id,
             },
+            accountIdByAddress: {
+              [mockOlderEvmAccount.address]: mockOlderEvmAccount.id,
+              [mockNewerEvmAccount.address]: mockNewerEvmAccount.id,
+              [mockBtcAccount.address]: mockBtcAccount.id,
+            },
           },
         });
 
@@ -2991,6 +3148,7 @@ describe('AccountsController', () => {
             accounts: {},
             selectedAccount: '',
           },
+          accountIdByAddress: {},
         },
       });
 
@@ -3037,6 +3195,11 @@ describe('AccountsController', () => {
               [mockNonEvmAccount.id]: mockNonEvmAccount,
             },
             selectedAccount: mockAccount.id,
+          },
+          accountIdByAddress: {
+            [mockAccount.address]: mockAccount.id,
+            [mockAccount2.address]: mockAccount2.id,
+            [mockNonEvmAccount.address]: mockNonEvmAccount.id,
           },
         },
       });
@@ -3134,6 +3297,15 @@ describe('AccountsController', () => {
             },
             selectedAccount: mockAccount.id,
           },
+          accountIdByAddress: {
+            [mockAccount.address]: mockAccount.id,
+            [mockAccount2.address]: mockAccount2.id,
+            [mockErc4337MainnetAccount.address]: mockErc4337MainnetAccount.id,
+            [mockErc4337TestnetAccount.address]: mockErc4337TestnetAccount.id,
+            [mockBtcMainnetAccount.address]: mockBtcMainnetAccount.id,
+            [mockBtcMainnetAccount2.address]: mockBtcMainnetAccount2.id,
+            [mockBtcTestnetAccount.address]: mockBtcTestnetAccount.id,
+          },
         },
       });
       expect(
@@ -3150,6 +3322,10 @@ describe('AccountsController', () => {
               [mockAccount2.id]: mockAccount2,
             },
             selectedAccount: mockAccount.id,
+          },
+          accountIdByAddress: {
+            [mockAccount.address]: mockAccount.id,
+            [mockAccount2.address]: mockAccount2.id,
           },
         },
       });
@@ -3173,6 +3349,10 @@ describe('AccountsController', () => {
               [mockAccount2.id]: mockAccount2,
             },
             selectedAccount: mockAccount.id,
+          },
+          accountIdByAddress: {
+            [mockAccount.address]: mockAccount.id,
+            [mockAccount2.address]: mockAccount2.id,
           },
         },
       });
@@ -3225,6 +3405,10 @@ describe('AccountsController', () => {
               },
               selectedAccount: mockAccount.id,
             },
+            accountIdByAddress: {
+              [mockAccount.address]: mockAccount.id,
+              [mockNonEvmAccount.address]: mockNonEvmAccount.id,
+            },
           },
         });
 
@@ -3258,6 +3442,9 @@ describe('AccountsController', () => {
           accounts: { [mockAccount.id]: mockAccount },
           selectedAccount: mockAccount.id,
         },
+        accountIdByAddress: {
+          [mockAccount.address]: mockAccount.id,
+        },
       },
     };
 
@@ -3286,6 +3473,10 @@ describe('AccountsController', () => {
               [mockAccount2.id]: mockAccount2,
             },
             selectedAccount: mockAccount.id,
+          },
+          accountIdByAddress: {
+            [mockAccount.address]: mockAccount.id,
+            [mockAccount2.address]: mockAccount2.id,
           },
         },
       });
@@ -3347,6 +3538,9 @@ describe('AccountsController', () => {
             accounts: { [mockAccount.id]: mockAccount },
             selectedAccount: mockAccount.id,
           },
+          accountIdByAddress: {
+            [mockAccount.address]: mockAccount.id,
+          },
         },
       });
       accountsController.setAccountName(mockAccount.id, 'new name');
@@ -3367,6 +3561,9 @@ describe('AccountsController', () => {
             accounts: { [mockAccount.id]: mockAccount },
             selectedAccount: mockAccount.id,
           },
+          accountIdByAddress: {
+            [mockAccount.address]: mockAccount.id,
+          },
         },
       });
 
@@ -3385,6 +3582,9 @@ describe('AccountsController', () => {
             internalAccounts: {
               accounts: { [mockAccount.id]: mockAccount },
               selectedAccount: mockAccount.id,
+            },
+            accountIdByAddress: {
+              [mockAccount.address]: mockAccount.id,
             },
           },
         });
@@ -3417,6 +3617,10 @@ describe('AccountsController', () => {
             },
             selectedAccount: mockAccount.id,
           },
+          accountIdByAddress: {
+            [mockAccount.address]: mockAccount.id,
+            [mockAccountWithName.address]: mockAccountWithName.id,
+          },
         },
       });
 
@@ -3433,6 +3637,9 @@ describe('AccountsController', () => {
           internalAccounts: {
             accounts: { [mockAccount.id]: mockAccount },
             selectedAccount: mockAccount.id,
+          },
+          accountIdByAddress: {
+            [mockAccount.address]: mockAccount.id,
           },
         },
       });
@@ -3511,6 +3718,9 @@ describe('AccountsController', () => {
             },
             selectedAccount: mockAccount.id,
           },
+          accountIdByAddress: {
+            [mockAccount.address]: mockAccount.id,
+          },
         },
         messenger,
       });
@@ -3549,6 +3759,9 @@ describe('AccountsController', () => {
               [mockAccount.id]: mockAccount,
             },
             selectedAccount: mockAccount.id,
+          },
+          accountIdByAddress: {
+            [mockAccount.address]: mockAccount.id,
           },
         },
         messenger,
@@ -3598,6 +3811,9 @@ describe('AccountsController', () => {
             accounts: { [mockAccount.id]: mockAccount },
             selectedAccount: mockAccount.id,
           },
+          accountIdByAddress: {
+            [mockAccount.address]: mockAccount.id,
+          },
         },
       });
 
@@ -3614,6 +3830,9 @@ describe('AccountsController', () => {
           internalAccounts: {
             accounts: { [mockAccount.id]: mockAccount },
             selectedAccount: mockAccount.id,
+          },
+          accountIdByAddress: {
+            [mockAccount.address]: mockAccount.id,
           },
         },
       });
@@ -3639,6 +3858,10 @@ describe('AccountsController', () => {
               [mockNonEvmAccount.id]: mockNonEvmAccount,
             },
             selectedAccount: mockAccount.id,
+          },
+          accountIdByAddress: {
+            [mockAccount.address]: mockAccount.id,
+            [mockNonEvmAccount.address]: mockNonEvmAccount.id,
           },
         },
       });
@@ -3676,6 +3899,9 @@ describe('AccountsController', () => {
               accounts: { [mockAccount.id]: mockAccount },
               selectedAccount: mockAccount.id,
             },
+            accountIdByAddress: {
+              [mockAccount.address]: mockAccount.id,
+            },
           },
           messenger,
         });
@@ -3706,6 +3932,10 @@ describe('AccountsController', () => {
               },
               selectedAccount: mockAccount.id,
             },
+            accountIdByAddress: {
+              [mockAccount.address]: mockAccount.id,
+              [mockNonEvmAccount.address]: mockNonEvmAccount.id,
+            },
           },
           messenger,
         });
@@ -3735,6 +3965,10 @@ describe('AccountsController', () => {
               },
               selectedAccount: mockAccount.id,
             },
+            accountIdByAddress: {
+              [mockAccount.address]: mockAccount.id,
+              [mockNonEvmAccount.address]: mockNonEvmAccount.id,
+            },
           },
           messenger,
         });
@@ -3757,6 +3991,9 @@ describe('AccountsController', () => {
             internalAccounts: {
               accounts: { [mockAccount.id]: mockAccount },
               selectedAccount: mockAccount.id,
+            },
+            accountIdByAddress: {
+              [mockAccount.address]: mockAccount.id,
             },
           },
           messenger,
@@ -3785,6 +4022,10 @@ describe('AccountsController', () => {
                 [mockAccount2.id]: mockAccount2,
               },
               selectedAccount: mockAccount.id,
+            },
+            accountIdByAddress: {
+              [mockAccount.address]: mockAccount.id,
+              [mockAccount2.address]: mockAccount2.id,
             },
           },
           messenger,
@@ -3823,6 +4064,9 @@ describe('AccountsController', () => {
               accounts: { [mockAccount.id]: mockAccount },
               selectedAccount: mockAccount.id,
             },
+            accountIdByAddress: {
+              [mockAccount.address]: mockAccount.id,
+            },
           },
           messenger,
         });
@@ -3841,6 +4085,9 @@ describe('AccountsController', () => {
             internalAccounts: {
               accounts: { [mockAccount.id]: mockAccount },
               selectedAccount: mockAccount.id,
+            },
+            accountIdByAddress: {
+              [mockAccount.address]: mockAccount.id,
             },
           },
           messenger,
@@ -3867,6 +4114,9 @@ describe('AccountsController', () => {
               accounts: { [mockAccount.id]: mockAccount },
               selectedAccount: mockAccount.id,
             },
+            accountIdByAddress: {
+              [mockAccount.address]: mockAccount.id,
+            },
           },
           messenger,
         });
@@ -3886,6 +4136,9 @@ describe('AccountsController', () => {
             internalAccounts: {
               accounts: { [mockAccount.id]: mockAccount },
               selectedAccount: mockAccount.id,
+            },
+            accountIdByAddress: {
+              [mockAccount.address]: mockAccount.id,
             },
           },
           messenger,
@@ -3965,6 +4218,7 @@ describe('AccountsController', () => {
         ),
       ).toMatchInlineSnapshot(`
         {
+          "accountIdByAddress": {},
           "internalAccounts": {
             "accounts": {},
             "selectedAccount": "",