+++ packages/accounts-controller/src/AccountsController.ts
@@ -43,6 +43,7 @@ import type { MultichainNetworkControllerNetworkDidChangeEvent } from './types';
 import type { AccountsControllerStrictState } from './typing';
 import type { HdSnapKeyringAccount } from './utils';
 import {
+  constructAccountIdByAddress,
   getEvmDerivationPathForIndex,
   getEvmGroupIndexFromAddressIndex,
   getUUIDFromAddressOfNormalAccount,
@@ -69,6 +70,7 @@ export type AccountsControllerState = {
     accounts: Record<AccountId, InternalAccount>;
     selectedAccount: string; // id of the selected account
   };
+  accountIdByAddress: Record<string, AccountId>;
 };
 
 /**
@@ -237,6 +239,12 @@ const accountsControllerMetadata = {
     includeInDebugSnapshot: false,
     usedInUi: true,
   },
+  accountIdByAddress: {
+    includeInStateLogs: false,
+    persist: false,
+    includeInDebugSnapshot: false,
+    usedInUi: true,
+  },
 };
 
 const defaultState: AccountsControllerState = {
@@ -244,6 +252,7 @@ const defaultState: AccountsControllerState = {
     accounts: {},
     selectedAccount: '',
   },
+  accountIdByAddress: {},
 };
 
 /**
@@ -308,6 +317,9 @@ export class AccountsController extends BaseController<
     messenger: AccountsControllerMessenger;
     state: AccountsControllerState;
   }) {
+    const accountIdByAddress = constructAccountIdByAddress(
+      state.internalAccounts.accounts,
+    );
     super({
       messenger,
       name: controllerName,
@@ -315,6 +327,7 @@ export class AccountsController extends BaseController<
       state: {
         ...defaultState,
         ...state,
+        accountIdByAddress,
       },
     });
 
@@ -477,9 +490,8 @@ export class AccountsController extends BaseController<
    * @returns The account with the specified address, or undefined if not found.
    */
   getAccountByAddress(address: string): InternalAccount | undefined {
-    return this.listMultichainAccounts().find(
-      (account) => account.address.toLowerCase() === address.toLowerCase(),
-    );
+    const accountId = this.state.accountIdByAddress[address.toLowerCase()];
+    return accountId ? this.getAccount(accountId) : undefined;
   }
 
   /**
@@ -667,6 +679,7 @@ export class AccountsController extends BaseController<
 
     this.#update((state) => {
       state.internalAccounts.accounts = internalAccounts;
+      state.accountIdByAddress = constructAccountIdByAddress(internalAccounts);
     });
   }
 
@@ -679,9 +692,13 @@ export class AccountsController extends BaseController<
    */
   loadBackup(backup: AccountsControllerState): void {
     if (backup.internalAccounts) {
+      const accountIdByAddress = constructAccountIdByAddress(
+        backup.internalAccounts.accounts,
+      );
       this.update(
         (currentState: WritableDraft<AccountsControllerStrictState>) => {
           currentState.internalAccounts = backup.internalAccounts;
+          currentState.accountIdByAddress = accountIdByAddress;
         },
       );
     }
@@ -904,11 +921,13 @@ export class AccountsController extends BaseController<
 
     this.#update(
       (state) => {
-        const { internalAccounts } = state;
+        const { internalAccounts, accountIdByAddress } = state;
 
         for (const patch of [patches.snap, patches.normal]) {
           for (const account of patch.removed) {
             delete internalAccounts.accounts[account.id];
+            // FIXME: We should not lowercase all addresses since some accounts addresses might be case-sensitive
+            delete accountIdByAddress[account.address.toLowerCase()];
 
             diff.removed.push(account.id);
           }
@@ -936,6 +955,9 @@ export class AccountsController extends BaseController<
                   lastSelected,
                 },
               };
+
+              // FIXME: We should not lowercase all addresses since some accounts addresses might be case-sensitive
+              accountIdByAddress[account.address.toLowerCase()] = account.id;
 
               diff.added.push(internalAccounts.accounts[account.id]);
             }